+++ src/main/kotlin/rs/kunpero/tacowizard/service/dto/FeedRequest.kt
@@ -1,4 +1,4 @@
-package rs.kunpero.fatpak.service.dto
+package rs.kunpero.tacowizard.service.dto
 
 class FeedRequestDto (
     val fromUser: String,
